requirements.txt
yt-dlp
moviepy
modal
whisperx
pyannote.audio
python-dotenv
openai
openpyxl
